테스트 로직별로 분리
