+++ src/test/java/com/waterfogsw/voucher/ConsoleTest.java
@@ -1,0 +1,33 @@
+package com.waterfogsw.voucher;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ConsoleTest {
+    //콘솔 입출력
+    @Test
+    @DisplayName("메뉴 입력")
+    public void inputMenu() {
+
+    }
+
+    @Test
+    @DisplayName("바우처 타입 입력")
+    public void inputVoucherType() {
+
+    }
+
+    @Test
+    @DisplayName("percent 입력")
+    public void inputPercent() {
+
+    }
+
+    @Test
+    @DisplayName("amount 입력")
+    public void inputAmount() {
+
+    }
+}
