+++ src/test/java/com/waterfogsw/voucher/VoucherFileRepositoryTests.java
@@ -1,0 +1,27 @@
+package com.waterfogsw.voucher;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class VoucherFileRepositoryTests {
+    //바우처 레포지토리
+    @Test
+    @DisplayName("파일 - 바우처 레포지토리에 저장")
+    public void fileRepositoryVoucherSave() {
+
+    }
+
+    @Test
+    @DisplayName("파일 - 바우처 리스트 조회")
+    public void fileRepositoryVoucherListLookUp() {
+
+    }
+
+    @Test
+    @DisplayName("파일 - 잘못된 경로")
+    public void fileRepositoryWrongPath() {
+
+    }
+}
