+++ src/test/java/com/waterfogsw/voucher/VoucherMemoryRepositoryTest.java
@@ -1,0 +1,21 @@
+package com.waterfogsw.voucher;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class VoucherMemoryRepositoryTest {
+    //바우처 레포지토리
+    @Test
+    @DisplayName("인메모리 - 바우처 레포지토리에 저장")
+    public void memoryRepositoryVoucherSave() {
+
+    }
+
+    @Test
+    @DisplayName("인메모리 - 바우처 리스트 조회")
+    public void memoryRepositoryVoucherListLookUp() {
+
+    }
+}
